% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area.circle}
\alias{area.circle}
\title{Calculate the area of a circle.}
\usage{
\method{area}{circle}(circ)
}
\arguments{
\item{circ}{An object of class 'shape' representing a circle.}
}
\value{
The area of the circle.
}
\description{
Calculate the area of a circle.
}
\examples{
# Create a circle object
circle <- list(radius = 3)
class(circle) <- "shape"

# Calculate the area of the circle
area.circle(circle)

}
\seealso{
\code{\link{area}}

Other area functions: 
\code{\link{area.cube}()},
\code{\link{area.cuboid}()},
\code{\link{area.cylinder}()},
\code{\link{area.rectangle}()},
\code{\link{area.sphere}()},
\code{\link{area.square}()},
\code{\link{area.triangle}()},
\code{\link{area}()}
}
\concept{area functions}
