% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area.cube}
\alias{area.cube}
\title{Calculate the surface area of a cube.}
\usage{
\method{area}{cube}(cub)
}
\arguments{
\item{cub}{An object of class 'shape' representing a cube with side length.}
}
\value{
The surface area of the cube.
}
\description{
Calculate the surface area of a cube.
}
\examples{
# Create a cube object
cube <- list(sideLength = 4)
class(cube) <- "shape"

# Calculate the surface area of the cube
area.cube(cube)

}
\seealso{
\code{\link{area}}

Other area functions: 
\code{\link{area.circle}()},
\code{\link{area.cuboid}()},
\code{\link{area.cylinder}()},
\code{\link{area.rectangle}()},
\code{\link{area.sphere}()},
\code{\link{area.square}()},
\code{\link{area.triangle}()},
\code{\link{area}()}
}
\concept{area functions}
