% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area.cylinder}
\alias{area.cylinder}
\title{Calculate the surface area of a cylinder.}
\usage{
\method{area}{cylinder}(shp)
}
\arguments{
\item{shp}{An object of class 'shape' representing a cylinder with radius and height.}
}
\value{
The surface area of the cylinder.
}
\description{
Calculate the surface area of a cylinder.
}
\examples{
\dontrun{
# Create a cylinder object
my_cylinder <- cylinder(radius = 2, height = 5)

# Calculate the surface area of the cylinder
area.cylinder(my_cylinder)
}
}
\seealso{
\code{\link{area}}

Other area functions: 
\code{\link{area.circle}()},
\code{\link{area.cube}()},
\code{\link{area.cuboid}()},
\code{\link{area.rectangle}()},
\code{\link{area.sphere}()},
\code{\link{area.square}()},
\code{\link{area.triangle}()},
\code{\link{area}()}
}
\concept{area functions}
