% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{Calculate the area of various 2D and 3D shapes.}
\usage{
area(shp)
}
\arguments{
\item{shp}{An object of class 'shape'.}
}
\description{
This provides functions to calculate the area of various 2D and 3D shapes,
specifically rectangles, squares, circles, triangles, cuboids, cubes, spheres, and cylinders.
The area can be calculated based on the dimensions of each shape.
}
\details{
The function checks if the input object is of class 'shape'. If not, it raises
an error indicating that the input must be an object of class 'shape'. If the input is
of class 'shape', it dispatches the calculation to the appropriate method based on the
type of shape.
}
\examples{
# Create a rectangle object
rectangle <- list(height = 4, width = 6)
class(rectangle) <- "shape"

# Calculate the area of the rectangle
area(rectangle)


}
\seealso{
Other area functions: 
\code{\link{area.circle}()},
\code{\link{area.cube}()},
\code{\link{area.cuboid}()},
\code{\link{area.cylinder}()},
\code{\link{area.rectangle}()},
\code{\link{area.sphere}()},
\code{\link{area.square}()},
\code{\link{area.triangle}()}
}
\concept{area functions}
