% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area.rectangle}
\alias{area.rectangle}
\title{Calculate the area of a rectangle.}
\usage{
\method{area}{rectangle}(shp)
}
\arguments{
\item{shp}{An object of class 'shape' representing a rectangle.}
}
\value{
The area of the rectangle.
}
\description{
Calculate the area of a rectangle.
}
\examples{
# Create a rectangle object
my_rectangle <- rectangle(height = 2,width = 4)

# Calculate the area of the shpangle
area.rectangle(my_rectangle)

}
\seealso{
\code{\link{area}}

Other area functions: 
\code{\link{area.circle}()},
\code{\link{area.cube}()},
\code{\link{area.cuboid}()},
\code{\link{area.cylinder}()},
\code{\link{area.sphere}()},
\code{\link{area.square}()},
\code{\link{area.triangle}()},
\code{\link{area}()}
}
\concept{area functions}
