% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area.sphere}
\alias{area.sphere}
\title{Calculate the surface area of a sphere.}
\usage{
\method{area}{sphere}(shp)
}
\arguments{
\item{shp}{An object of class 'shape' representing a shpere with radius.}
}
\value{
The surface area of the shpere.
}
\description{
Calculate the surface area of a sphere.
}
\examples{
\dontrun{
# Create a shpere object
my_shpere <- shpere(radius = 2)

# Calculate the surface area of the shpere
area.shpere(my_shpere)
}
}
\seealso{
\code{\link{area}}

Other area functions: 
\code{\link{area.circle}()},
\code{\link{area.cube}()},
\code{\link{area.cuboid}()},
\code{\link{area.cylinder}()},
\code{\link{area.rectangle}()},
\code{\link{area.square}()},
\code{\link{area.triangle}()},
\code{\link{area}()}
}
\concept{area functions}
