% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area.square}
\alias{area.square}
\title{Calculate the area of a square.}
\usage{
\method{area}{square}(shp)
}
\arguments{
\item{shp}{An object of class 'shape' representing a shpare.}
}
\value{
The area of the shpare.
}
\description{
Calculate the area of a square.
}
\examples{
\dontrun{
# Create a shpare object
my_shpare1 <- shpare(sideLength = 5)


# Calculate the area of the shpare
area.shpare(my_shpare1)
}

}
\seealso{
\code{\link{area}}

Other area functions: 
\code{\link{area.circle}()},
\code{\link{area.cube}()},
\code{\link{area.cuboid}()},
\code{\link{area.cylinder}()},
\code{\link{area.rectangle}()},
\code{\link{area.sphere}()},
\code{\link{area.triangle}()},
\code{\link{area}()}
}
\concept{area functions}
