% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{circle}
\alias{circle}
\title{Create a circle object.}
\usage{
circle(radius)
}
\arguments{
\item{radius}{Radius of the circle.}
}
\value{
An object of class 'shape', '2dShape', and 'circle'.
}
\description{
Create a circle object.
}
\examples{
my_circle <- circle(radius = 3)

}
\seealso{
Other 2D Shape Classes: 
\code{\link{rectangle}()},
\code{\link{square}()},
\code{\link{triangle}()}
}
\concept{2D Shape Classes}
