% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{cube}
\alias{cube}
\title{Create a cube object.}
\usage{
cube(sideLength)
}
\arguments{
\item{sideLength}{Side length of the cube.}
}
\value{
An object of class 'shape', '3dShape', and 'cube'.
}
\description{
Create a cube object.
}
\examples{
\dontrun{
my_cube1 <- cube(sideLength = 5)
my_cube2 <- cuboid(height = 3, width = 3, depth = 3)
}
}
\seealso{
\code{\link{cuboid}}

Other 3D Shape Classes: 
\code{\link{cuboid}()},
\code{\link{cylinder}()},
\code{\link{sphere}()}
}
\concept{3D Shape Classes}
