% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{cuboid}
\alias{cuboid}
\title{Create a cuboid object.}
\usage{
cuboid(height, width, depth)
}
\arguments{
\item{height}{Height of the cuboid.}

\item{width}{Width of the cuboid.}

\item{depth}{Depth of the cuboid.}
}
\value{
An object of class 'shape', '3dShape', and 'cuboid'.
}
\description{
Create a cuboid object.
}
\examples{
my_cuboid1 <- cuboid(height = 3, width = 4, depth = 5)


}
\seealso{
Other 3D Shape Classes: 
\code{\link{cube}()},
\code{\link{cylinder}()},
\code{\link{sphere}()}
}
\concept{3D Shape Classes}
