% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{cylinder}
\alias{cylinder}
\title{Create a cylinder object.}
\usage{
cylinder(radius, height)
}
\arguments{
\item{radius}{Radius of the cylinder.}

\item{height}{Height of the cylinder.}
}
\value{
An object of class 'shape', '3dShape', and 'cylinder'.
}
\description{
Create a cylinder object.
}
\examples{
\dontrun{
my_cylinder <- cylinder(radius = 2, height = 5)
}
}
\seealso{
Other 3D Shape Classes: 
\code{\link{cube}()},
\code{\link{cuboid}()},
\code{\link{sphere}()}
}
\concept{3D Shape Classes}
