% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.circle}
\alias{draw.circle}
\title{Draw a circle in 2D.}
\usage{
\method{draw}{circle}(shape, color = "blue", fillColor = "transparent")
}
\arguments{
\item{shape}{An object of class 'shape' representing a circle.}

\item{color}{The border color of the circle (default is "blue").}

\item{fillColor}{The fill color of the circle (default is "transparent").}
}
\description{
Draw a circle in 2D.
}
\examples{
# Create a circle
my_circle <- circle(radius = 3)

# Draw the circle
draw.circle(my_circle)

}
\seealso{
Other draw functions: 
\code{\link{draw.cube}()},
\code{\link{draw.cuboid}()},
\code{\link{draw.cylinder}()},
\code{\link{draw.rectangle}()},
\code{\link{draw.sphere}()},
\code{\link{draw.square}()},
\code{\link{draw.triangle}()},
\code{\link{draw}()}
}
\concept{draw functions}
