% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.cube}
\alias{draw.cube}
\title{Draw a cube in 3D using rgl.}
\usage{
\method{draw}{cube}(shape, color = "red", ...)
}
\arguments{
\item{shape}{An object of class 'shape' representing a cube.}

\item{color}{The color of the cube (default is "red").}

\item{...}{Additional arguments passed to underlying rgl functions.}
}
\description{
Draw a cube in 3D using rgl.
}
\examples{
\dontrun{
# Create a cube
my_cube <- cube(sideLength = 5)

# Draw the cube
draw.cube(my_cube)
}
}
\seealso{
Other draw functions: 
\code{\link{draw.circle}()},
\code{\link{draw.cuboid}()},
\code{\link{draw.cylinder}()},
\code{\link{draw.rectangle}()},
\code{\link{draw.sphere}()},
\code{\link{draw.square}()},
\code{\link{draw.triangle}()},
\code{\link{draw}()}
}
\concept{draw functions}
