% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.cuboid}
\alias{draw.cuboid}
\title{Draw a cuboid in 3D using rgl.}
\usage{
\method{draw}{cuboid}(shape, color = "red", ...)
}
\arguments{
\item{shape}{An object of class 'shape' representing a cube.}

\item{color}{The color of the cube (default is "red").}
}
\description{
Draw a cuboid in 3D using rgl.
}
\examples{
# Create a cube
my_cuboid <- cuboid(height = 3, width = 3, depth = 3)

# Draw the cube
draw.cuboid(my_cuboid)

}
\seealso{
Other draw functions: 
\code{\link{draw.circle}()},
\code{\link{draw.cube}()},
\code{\link{draw.cylinder}()},
\code{\link{draw.rectangle}()},
\code{\link{draw.sphere}()},
\code{\link{draw.square}()},
\code{\link{draw.triangle}()},
\code{\link{draw}()}
}
\concept{draw functions}
