% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.cylinder}
\alias{draw.cylinder}
\title{Draw a cylinder in 3D.}
\usage{
\method{draw}{cylinder}(shape, color = "red", ...)
}
\arguments{
\item{shape}{An object of class 'shape' representing a cylinder.}

\item{color}{The color of the cylinder (default is "red").}

\item{...}{Additional arguments passed to `cylinder3d` and
other underlying functions.}
}
\description{
Draw a cylinder in 3D.
}
\examples{
# Create a cylinder
my_cylinder <- cylinder(radius = 2, height = 5)

# Draw the cylinder
draw.cylinder(my_cylinder)

}
\seealso{
Other draw functions: 
\code{\link{draw.circle}()},
\code{\link{draw.cube}()},
\code{\link{draw.cuboid}()},
\code{\link{draw.rectangle}()},
\code{\link{draw.sphere}()},
\code{\link{draw.square}()},
\code{\link{draw.triangle}()},
\code{\link{draw}()}
}
\concept{draw functions}
