% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw/Visualize shapes in 2D and 3D.}
\usage{
draw(shape, color = "blue", fillColor = "transparent")
}
\arguments{
\item{shape}{An object of class 'shape'.}

\item{color}{The border color of the shape (default is "blue" for 2D shapes
and "red" for 3d shapes but can be easily changed.).}

\item{fillColor}{The fill color of the shape (default is "transparent" but
can be easily changed).}
}
\description{
This  provides functions to visually draw various shapes,
including 2D shapes such rectangles, squares, circles, triangles,
and 3D shapes such as cuboids, cubes, spheres, and cylinders.
}
\details{
The function checks if the input object is of class 'shape'. If not,
it raises an error indicating that the input must be an object of class
'shape'.If the input is of class 'shape', it dispatches the drawing to the
 appropriate method based on the type of shape.
}
\examples{
\dontrun{
# Create a rectangle
my_rectangle <- rectangle(height = 4, width = 6)

#draw the rectangle
draw(my_rectangle)

# Create a sphere
my_sphere <- sphere(radius = 2)

# Draw the sphere
draw(my_sphere)
}

}
\seealso{
Other draw functions: 
\code{\link{draw.circle}()},
\code{\link{draw.cube}()},
\code{\link{draw.cuboid}()},
\code{\link{draw.cylinder}()},
\code{\link{draw.rectangle}()},
\code{\link{draw.sphere}()},
\code{\link{draw.square}()},
\code{\link{draw.triangle}()}
}
\concept{draw functions}
