% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.rectangle}
\alias{draw.rectangle}
\title{Draw a rectangle in 2D.}
\usage{
\method{draw}{rectangle}(shape, color = "blue", fillColor = "transparent")
}
\arguments{
\item{shape}{An object of class 'shape' representing a rectangle.}

\item{color}{The border color of the rectangle (default is "blue").}

\item{fillColor}{The fill color of the rectangle (default is "transparent").}
}
\description{
Draw a rectangle in 2D.
}
\examples{
# Create a rectangle
my_rectangle <- rectangle(height = 4, width = 6)

# Draw the rectangle
draw.rectangle(my_rectangle)

}
\seealso{
\code{\link{draw}}

Other draw functions: 
\code{\link{draw.circle}()},
\code{\link{draw.cube}()},
\code{\link{draw.cuboid}()},
\code{\link{draw.cylinder}()},
\code{\link{draw.sphere}()},
\code{\link{draw.square}()},
\code{\link{draw.triangle}()},
\code{\link{draw}()}
}
\concept{draw functions}
