% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.sphere}
\alias{draw.sphere}
\title{Draw a sphere in 3D.}
\usage{
\method{draw}{sphere}(sph, color = "#FF0000", ...)
}
\arguments{
\item{color}{The color of the sphere (default is "#FF0000 or 'red').}

\item{...}{Additional arguments passed to `sphere3d` and other underlying functions.}

\item{shape}{An object of class 'shape' representing a sphere.}
}
\description{
Draw a sphere in 3D.
}
\examples{
# Create a sphere
my_sphere <- sphere(radius = 2)

# Draw the sphere
draw.sphere(my_sphere)

}
\seealso{
Other draw functions: 
\code{\link{draw.circle}()},
\code{\link{draw.cube}()},
\code{\link{draw.cuboid}()},
\code{\link{draw.cylinder}()},
\code{\link{draw.rectangle}()},
\code{\link{draw.square}()},
\code{\link{draw.triangle}()},
\code{\link{draw}()}
}
\concept{draw functions}
