% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.square}
\alias{draw.square}
\title{Draw a square in 2D.}
\usage{
\method{draw}{square}(shape, color = "blue", fillColor = "transparent")
}
\arguments{
\item{shape}{An object of class 'shape' representing a square.}

\item{color}{The border color of the square (default is "blue").}

\item{fillColor}{The fill color of the square (default is "transparent").}
}
\description{
Draw a square in 2D.
}
\examples{
# Create a square
my_square <- square(sideLength = 5)

# Draw the square
draw.square(my_square)

}
\seealso{
\code{\link{draw}}

Other draw functions: 
\code{\link{draw.circle}()},
\code{\link{draw.cube}()},
\code{\link{draw.cuboid}()},
\code{\link{draw.cylinder}()},
\code{\link{draw.rectangle}()},
\code{\link{draw.sphere}()},
\code{\link{draw.triangle}()},
\code{\link{draw}()}
}
\concept{draw functions}
