% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.triangle}
\alias{draw.triangle}
\title{Draw a triangle in 2D.}
\usage{
\method{draw}{triangle}(shape, color = "blue", fillColor = "transparent")
}
\arguments{
\item{shape}{An object of class 'shape' representing a triangle.}

\item{color}{The border color of the triangle (default is "blue").}

\item{fillColor}{The fill color of the triangle (default is "transparent").}
}
\description{
Draw a triangle in 2D.
}
\examples{
# Create a triangle
my_triangle <- triangle(a = 3, b = 4, c = 5)

# Draw the triangle
draw.triangle(my_triangle)

}
\seealso{
Other draw functions: 
\code{\link{draw.circle}()},
\code{\link{draw.cube}()},
\code{\link{draw.cuboid}()},
\code{\link{draw.cylinder}()},
\code{\link{draw.rectangle}()},
\code{\link{draw.sphere}()},
\code{\link{draw.square}()},
\code{\link{draw}()}
}
\concept{draw functions}
