% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_list_helpers.R
\name{expand_rdf_list}
\alias{expand_rdf_list}
\title{Expand an RDF list into an ordered vector}
\usage{
expand_rdf_list(head, triples)
}
\arguments{
\item{head}{Identifier of the head node for the RDF list.}

\item{triples}{Data frame of RDF triples with columns subject, predicate, object.}
}
\value{
Character vector of list members in order. If the head does not look
like an RDF list node, the head value itself is returned.
}
\description{
Expand an RDF list into an ordered vector
}
\examples{
triples <- data.frame(
  subject = c("_:b0", "_:b0", "_:b1", "_:b1"),
  predicate = c(
    "<http://www.w3.org/1999/02/22-rdf-syntax-ns#first>",
    "<http://www.w3.org/1999/02/22-rdf-syntax-ns#rest>",
    "<http://www.w3.org/1999/02/22-rdf-syntax-ns#first>",
    "<http://www.w3.org/1999/02/22-rdf-syntax-ns#rest>"
  ),
  object = c("ex:first", "_:b1", "ex:second", "<http://www.w3.org/1999/02/22-rdf-syntax-ns#nil>"),
  stringsAsFactors = FALSE
)

expand_rdf_list("_:b0", triples)
}
