% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_or_na.R
\name{first_or_na}
\alias{first_or_na}
\title{Return the first element or NA when empty.}
\usage{
first_or_na(x)
}
\arguments{
\item{x}{A vector or list.}
}
\value{
The first element of \code{x}, or \code{NA} if \code{x} is empty.
}
\description{
Return the first element or NA when empty.
}
\examples{
first_or_na(c("a", "b"))
first_or_na(list())
}
