% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeR.R
\docType{data}
\name{FISH}
\alias{FISH}
\title{An example data file}
\format{An example data file}
\source{
\url{https://github.com/lisalibungan/shapeR}
}
\usage{
data(FISH)
}
\description{
The file's columns are:
\itemize{
  \item country
  \item station
  \item pop
  \item stockID
  \item day
  \item month
  \item year
  \item lat
  \item lon
  \item fishno
  \item length_cm
  \item weight_g
  \item age
  \item sex
  \item maturity
  \item folder
  \item picname
  \item cal
}
}
\keyword{datasets}

