% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeR.R
\name{getFourier}
\alias{getFourier}
\title{Get Fourier coefficients, filtered according to filter}
\usage{
getFourier(object)
}
\arguments{
\item{object}{\code{\linkS4class{shapeR}} object}
}
\value{
The Fourier coefficients for all fish as determined by \code{\link{setFilter}}
}
\description{
Returns the Fourier coefficients determined in \code{\link{stdCoefs}}.
Returns only values as set in \code{\link{setFilter}}
}
\author{
Lisa Anne Libungan
}

