% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeR.R
\name{getMasterlist}
\alias{getMasterlist}
\title{Get filtered master.list values}
\usage{
getMasterlist(object, useFilter = TRUE)
}
\arguments{
\item{object}{\code{\linkS4class{shapeR}} object}

\item{useFilter}{If TRUE, the master.list values are filtered by the slot \code{filter}. FALSE = no filtering.}
}
\value{
The \code{master.list} is filtered by the slot \code{filter} if the \code{useFilter} is TRUE, else no filtering is done.
}
\description{
Returns selected values from \code{master.list}
}
\author{
Lisa Anne Libungan
}

