% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeR.R
\name{getStdFourier}
\alias{getStdFourier}
\title{Get standardized Fourier coefficients, filtered according to filter}
\usage{
getStdFourier(object)
}
\arguments{
\item{object}{\code{\linkS4class{shapeR}} object}
}
\value{
The standardized Fourier coefficients for all valid fish as determined by the slot \code{filter}
}
\description{
Returns the standardized Fourier coefficients determined in  \code{\link{stdCoefs}}.
Returns only values as set in the slot \code{filter}
}
\author{
Lisa Anne Libungan
}

