% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeR.R
\name{getStdMeasurements}
\alias{getStdMeasurements}
\title{Get simple shape variables after standardization, filtered according to filter}
\usage{
getStdMeasurements(object)
}
\arguments{
\item{object}{\code{\linkS4class{shapeR}} object}
}
\value{
A data frame with all valid fish as determined by the slot \code{filter}. Returns only variables that have not been removed after standardization.
}
\description{
Returns the simple shape variables determined in \code{stdCoefs}.
Returns only values as set in the slot \code{filter}
}
\examples{
data(otoliths)
#Calculate the mean standardized otolith length for each fish population
tapply(getStdMeasurements(shape)$otolith.length,
getMasterlist(shape)$pop,mean)
}
\author{
Lisa Anne Libungan
}

