% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeR.R
\name{getStdWavelet}
\alias{getStdWavelet}
\title{Get standardized wavelet coefficients, filtered according to filter}
\usage{
getStdWavelet(object)
}
\arguments{
\item{object}{\code{\linkS4class{shapeR}} object}
}
\value{
The standardized wavelet coefficients for all valid fish as determined by the slot \code{filter}
}
\description{
Returns the standardized wavelet coefficients determined in  \code{stdCoefs}.
Returns only values as set in the slot \code{filter}
}
\author{
Lisa Anne Libungan
}

