% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeR.R
\name{getWavelet}
\alias{getWavelet}
\title{Get wavelet coefficients, filtered according to filter}
\usage{
getWavelet(object)
}
\arguments{
\item{object}{\code{\linkS4class{shapeR}} object}
}
\value{
The wavelet coefficients for all valid fish as determined by the slot \code{filter}
}
\description{
Returns the wavelet coefficients determined in \code{\link{generateShapeCoefficients}}.
Returns only values as set in the slot \code{filter}
}
\author{
Lisa Anne Libungan
}

