% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{is.numeric.value}
\alias{is.numeric.value}
\title{Function to check if the input is numeric.}
\usage{
\method{is.numeric}{value}(x)
}
\arguments{
\item{x}{Input value to be checked.}
}
\value{
Logical value indicating whether the input is numeric.
}
\description{
Function to check if the input is numeric.
}
