% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-iri.R
\name{is_curie}
\alias{is_curie}
\title{Identify CURIE-like identifiers}
\usage{
is_curie(x)
}
\arguments{
\item{x}{Character vector to inspect.}
}
\value{
Logical vector marking elements that look like CURIEs.
}
\description{
Identify CURIE-like identifiers
}
\examples{
is_curie(c("ex:Person", "http://example.com/Person", NA))
}
