% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_list_helpers.R
\name{is_rdf_nil}
\alias{is_rdf_nil}
\title{Determine whether a node is rdf:nil}
\usage{
is_rdf_nil(node)
}
\arguments{
\item{node}{Node identifier string.}
}
\description{
Determine whether a node is rdf:nil
}
\keyword{internal}
