% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimeter.R
\name{perimeter.circle}
\alias{perimeter.circle}
\title{Calculate the perimeter of a circle.}
\usage{
\method{perimeter}{circle}(shp)
}
\arguments{
\item{shp}{An object of class '2dShape' representing a circle.}
}
\value{
The perimeter of the circle.
}
\description{
Calculate the perimeter of a circle.
}
\examples{
\dontrun{
# Create a circle
my_circle <- circle(radius = 3)

# Calculate the perimeter of the circle
perimeter.circle(my_circle)
}
}
\seealso{
\code{\link{perimeter}}

Other perimeter functions: 
\code{\link{perimeter.rectangle}()},
\code{\link{perimeter.square}()},
\code{\link{perimeter.triangle}()},
\code{\link{perimeter}()}
}
\concept{perimeter functions}
