% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimeter.R
\name{perimeter}
\alias{perimeter}
\title{Calculate the perimeter of 2D shapes.}
\usage{
perimeter(shp)
}
\arguments{
\item{shp}{An object of class '2dShape'.}
}
\description{
This package provides functions to calculate the perimeter of various 2D shapes,
including rectangles, squares, circles, and triangles. The perimeter can be
calculated based on the dimensions of each shape.
}
\details{
The function checks if the input object is of class '2dShape'. If not,
it raises an error indicating that the input must be an object of class '2dShape'.
If the input is of class '2dShape', it dispatches the calculation to the appropriate
method based on the type of shape.
}
\seealso{
\code{\link{perimeter.rectangle}}, \code{\link{perimeter.square}},
\code{\link{perimeter.circle}}, \code{\link{perimeter.triangle}}

Other perimeter functions: 
\code{\link{perimeter.circle}()},
\code{\link{perimeter.rectangle}()},
\code{\link{perimeter.square}()},
\code{\link{perimeter.triangle}()}
}
\concept{perimeter functions}
