% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimeter.R
\name{perimeter.rectangle}
\alias{perimeter.rectangle}
\title{Calculate the perimeter of a rectangle.}
\usage{
\method{perimeter}{rectangle}(shp)
}
\arguments{
\item{shp}{An object of class '2dShape' representing a rectangle.}
}
\value{
The perimeter of the rectangle.
}
\description{
Calculate the perimeter of a rectangle.
}
\examples{
\dontrun{
# Create a rectangle
my_rectangle <- rectangle(height = 2, width = 4)

# Calculate the perimeter of the rectangle
perimeter.rectangle(my_rectangle)
}

}
\seealso{
\code{\link{perimeter}}

Other perimeter functions: 
\code{\link{perimeter.circle}()},
\code{\link{perimeter.square}()},
\code{\link{perimeter.triangle}()},
\code{\link{perimeter}()}
}
\concept{perimeter functions}
