% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimeter.R
\name{perimeter.square}
\alias{perimeter.square}
\title{Calculate the perimeter of a square.}
\usage{
\method{perimeter}{square}(shp)
}
\arguments{
\item{shp}{An object of class '2dShape' representing a square.}
}
\value{
The perimeter of the square.
}
\description{
Calculate the perimeter of a square.
}
\examples{
# Create a square
my_square <- square(sideLength = 5)

# Calculate the perimeter of the square
perimeter.square(my_square)

}
\seealso{
\code{\link{perimeter}}

Other perimeter functions: 
\code{\link{perimeter.circle}()},
\code{\link{perimeter.rectangle}()},
\code{\link{perimeter.triangle}()},
\code{\link{perimeter}()}
}
\concept{perimeter functions}
