% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimeter.R
\name{perimeter.triangle}
\alias{perimeter.triangle}
\title{Calculate the perimeter of a triangle.}
\usage{
\method{perimeter}{triangle}(tri)
}
\arguments{
\item{tri}{An object of class '2dShape' representing a triangle.}
}
\value{
The perimeter of the triangle.
}
\description{
Calculate the perimeter of a triangle.
}
\examples{
# Create a triangle
my_triangle <- triangle(a = 3, b = 4, c = 5)

# Calculate the perimeter of the triangle
perimeter.triangle(my_triangle)

}
\seealso{
\code{\link{perimeter}}

Other perimeter functions: 
\code{\link{perimeter.circle}()},
\code{\link{perimeter.rectangle}()},
\code{\link{perimeter.square}()},
\code{\link{perimeter}()}
}
\concept{perimeter functions}
