% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeR.R
\name{plotFourierShape}
\alias{plotFourierShape}
\title{Mean otolith shape based on Fourier reconstruction}
\usage{
plotFourierShape(object, class.name, show.angle = FALSE,lty=1:5,col=1:6, ...)
}
\arguments{
\item{object}{A \code{\linkS4class{shapeR}} object}

\item{class.name}{A string as the column name in the master list}

\item{show.angle}{If TRUE angles are shown on the plot}

\item{lty,col}{Vector of line types and colors. Values are used cyclically.}

\item{...}{Additional parameters to be passed to 'plot'}
}
\description{
A function for showing the mean otolith shape based on Fourier reconstruction
}
\examples{
data(shape)
plotFourierShape(shape, "pop",show.angle = TRUE,lwd=2,lty=1)
}
\author{
Lisa Anne Libungan
}
\references{
Libungan LA and Palsson S (2015) ShapeR: An R Package to Study Otolith Shape Variation among Fish Populations. PLoS ONE 10(3): e0121102. \url{http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0121102}
}

