% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_list_helpers.R
\name{rdf_list_constants}
\alias{rdf_list_constants}
\title{Helpers for working with RDF lists.}
\usage{
rdf_list_constants()
}
\description{
Functions to traverse RDF collections represented with rdf:first / rdf:rest
triples and return the contained values in order.
}
\keyword{internal}
