% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_list_helpers.R
\name{rdf_list_rest_value}
\alias{rdf_list_rest_value}
\title{Get the rdf:rest value for a given list node}
\usage{
rdf_list_rest_value(node, triples)
}
\arguments{
\item{node}{List node identifier.}

\item{triples}{Data frame of RDF triples with columns subject, predicate, object.}
}
\description{
Get the rdf:rest value for a given list node
}
\keyword{internal}
