% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeR.R
\name{read.master.list}
\alias{read.master.list}
\title{Read updated master list}
\usage{
read.master.list(object, ...)
}
\arguments{
\item{object}{A \code{\linkS4class{shapeR}} object}

\item{...}{Additional parameter for \code{read.csv} for reading the info.file}
}
\value{
\code{\linkS4class{shapeR}} object with values in slots:
\itemize{
  \item master.list.org
}
}
\description{
Reads an updated master list. This is important to run if you want to ensure that a updated master list is used in the analysis.
}
\author{
Lisa Anne Libungan
}

