% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{rectangle}
\alias{rectangle}
\title{Create a rectangle object.}
\usage{
rectangle(height, width)
}
\arguments{
\item{height}{Height of the rectangle.}

\item{width}{Width of the rectangle.}
}
\value{
An object of class 'shape', '2dShape', and 'rectangle'.
}
\description{
Create a rectangle object.
}
\examples{
my_rectangle <- rectangle(height = 2,width = 4)

}
\seealso{
Other 2D Shape Classes: 
\code{\link{circle}()},
\code{\link{square}()},
\code{\link{triangle}()}
}
\concept{2D Shape Classes}
