% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeR.R
\docType{data}
\name{shape}
\alias{shape}
\title{An example shapeR instance including 160 images.
The shape coefficients have been generated.
The wavelet coefficients have been standardized using pop and length_cm.}
\format{A shapeR class including 160 images}
\usage{
data(otoliths)
}
\description{
The class slot's are as follows:
\itemize{
  \item project.path. A path as "ShapeAnalysis/"
  \item info.file. A file as FISH.csv. The information is stored in the data frame master.list
  \item outline.list. A list with three elements (IC, NO, SC) which give a list of the otolith outlines
  \item filter. A logical vector showing which elements of the master list have valid otoliths
  \item fourier.coef. A matrix of the Normalized Elliptic Fourier coefficients
  \item wavelet.coef. A matrix of the wavelet coefficients
  \item shape. A matrix of shape variables after scaling according to calibration otolith.area, otolith.length, otolith.width, otolith.perimeter.
  \item fourier.coef.std. A matrix which will contain standardized Fourier coefficients
  \item wavelet.coef.std. A matrix which will contain standardized wavelet coefficients
  \item shape.coef.raw. A matrix of shape variables before scaling according to calibration otolith.area, otolith.length, otolith.width, otolith.perimeter.
  \item master.list. The contents of the info.file
}
}
\keyword{datasets}

