% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaper.R
\name{shaper_country}
\alias{shaper_country}
\title{shaper_country}
\usage{
shaper_country(country)
}
\arguments{
\item{country}{The name of the country.}
}
\value{
Country's ISO code.
}
\description{
This function allows you to find and search for the right country's ISO code.
If no argument is filed, each ISO code of each country will be displayed.
}
\examples{
mycountry <- shaper_country()
mycountry <- shaper_country(country = "Italy")
mycountry <- shaper_country("Italy")

}
\seealso{
\code{\link{shaper_data}} to collect shapefiles when you have the country code.
}
