% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaper.R
\name{shaper_data}
\alias{shaper_data}
\title{shaper_data}
\usage{
shaper_data(country_code, level = 0)
}
\arguments{
\item{country_code}{The country's ISO code.}

\item{level}{The shapefile's level of granularity.}
}
\value{
Shapefile for the country and level requested.
}
\description{
This function allows you to retrive the wanted shapefile.
}
\examples{
map <- shaper_data(country_code = "ITA")
map <- shaper_data(country_code = "ITA", level=1)
map <- shaper_data("ITA", 1)
map <- shaper_data("USA", 2)

}
\seealso{
\code{\link{shaper_country}} for the country's ISO code.
}
