% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_iri.R
\name{shorten_iri}
\alias{shorten_iri}
\title{Shorten IRIs using known prefixes.}
\usage{
shorten_iri(iri, prefixes = character())
}
\arguments{
\item{iri}{Character vector of full IRIs.}

\item{prefixes}{Named character vector mapping prefix to base IRI. Optional;
defaults to an empty vector meaning no shortening will be applied.}
}
\value{
Character vector of compact IRIs (e.g., \code{ex:Resource}).
}
\description{
Shorten IRIs using known prefixes.
}
\examples{
prefixes <- c(ex = "http://example.com/")
shorten_iri("http://example.com/Resource", prefixes)
shorten_iri(c("http://example.com/id/1", "http://other.com/id/1"), prefixes)
}
