% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeR.R
\name{show.original.with.outline}
\alias{show.original.with.outline}
\title{Show the extracted outline on top of the original image}
\usage{
show.original.with.outline(object, folder, fname)
}
\arguments{
\item{object}{A \code{\linkS4class{shapeR}} object}

\item{folder}{The folder name where the image is stored}

\item{fname}{Image file name. Not including the extension ".jpg"}
}
\description{
A function which displayes the outlines which were extracted from the image in the "Fixed" folder on top of the corresponding image in the "Original" folder.
}
\examples{
\dontrun{
#Follow the example in Libungan and Palsson (2015) and run the following lines:
show.original.with.outline(shape,"IC","403_2")}
}
\author{
Lisa Anne Libungan
}
\references{
Libungan LA and Palsson S (2015) ShapeR: An R Package to Study Otolith Shape Variation among Fish Populations. PLoS ONE 10(3): e0121102.
}

