% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeR.R
\docType{methods}
\name{show,shapeR-method}
\alias{show,shapeR-method}
\title{Show a shapeR object}
\usage{
\S4method{show}{shapeR}(object)
}
\arguments{
\item{object}{a shapeR oject}
}
\description{
Show the project, the number of outlines that have been read and which fundamental methods have been run.
}
