% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{sphere}
\alias{sphere}
\title{Create a sphere object.}
\usage{
sphere(radius)
}
\arguments{
\item{radius}{Radius of the sphere.}
}
\value{
An object of class 'shape', '3dShape', and 'sphere'.
}
\description{
Create a sphere object.
}
\examples{
my_sphere <- sphere(radius = 2)

}
\seealso{
Other 3D Shape Classes: 
\code{\link{cube}()},
\code{\link{cuboid}()},
\code{\link{cylinder}()}
}
\concept{3D Shape Classes}
