% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{square}
\alias{square}
\title{Create a square object.}
\usage{
square(sideLength)
}
\arguments{
\item{sideLength}{Side length of the square.}
}
\value{
An object of class 'shape', '2dShape', and 'square'.
}
\description{
Create a square object.
}
\examples{
\dontrun{
my_square1 <- square(sideLength = 5)
my_square2 < rectangle(height = 5, width = 5)
}
}
\seealso{
\code{\link{rectangle}}

Other 2D Shape Classes: 
\code{\link{circle}()},
\code{\link{rectangle}()},
\code{\link{triangle}()}
}
\concept{2D Shape Classes}
