% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.2dShape}
\alias{summary.2dShape}
\title{Generate a summary of the parameters for 2D shapes.}
\usage{
\method{summary}{`2dShape`}(shp)
}
\arguments{
\item{shp}{An object of class 'shape' representing a 2D shape.}
}
\value{
Name of the shape, Area and Perimeter
}
\description{
Generate a summary of the parameters for 2D shapes.
}
\examples{
\dontrun{
my_square <- square(sideLength = 5)

# Generate a summary for the square
summary.2dShape(my_square)
}
}
\seealso{
Other summary functions: 
\code{\link{summary.3dShape}()},
\code{\link{summary}()}
}
\concept{summary functions}
