% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.3dShape}
\alias{summary.3dShape}
\title{Generate a summary of the parameters for 3D shapes.}
\usage{
\method{summary}{`3dShape`}(shp)
}
\arguments{
\item{shp}{An object of class 'shape' representing a 3D shape.}
}
\description{
Generate a summary of the parameters for 3D shapes.
}
\examples{
# Create a cylinder
my_cylinder1 <- cylinder(radius = 2, height = 5)

# Generate a summary for the cylinder
summary.3dShape(my_cylinder1)

}
\seealso{
Other summary functions: 
\code{\link{summary.2dShape}()},
\code{\link{summary}()}
}
\concept{summary functions}
