% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Generate a summary for shapes.}
\usage{
summary(shp)
}
\arguments{
\item{shp}{An object of class 'shape'.}
}
\description{
This function provides a summary for objects of class 'shape'. The summary
includes information such as the shape's class, perimeter (for 2D shapes),
surface area (for 3D shapes), and volume (for 3D shapes).
}
\details{
The function checks if the input object is of class 'shape'. If not,
it raises an error indicating that the input must be an object of class 'shape'.
If the input is of class 'shape', it dispatches the summary calculation to the
appropriate method based on the type of shape.
}
\seealso{
Other summary functions: 
\code{\link{summary.2dShape}()},
\code{\link{summary.3dShape}()}
}
\concept{summary functions}
