% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{triangle}
\alias{triangle}
\title{Create a triangle object.}
\usage{
triangle(a, b, c)
}
\arguments{
\item{a}{Length of side a.}

\item{b}{Length of side b.}

\item{c}{Length of side c.}
}
\value{
An object of class 'shape', '2dShape', and 'triangle'.
}
\description{
Create a triangle object.
}
\examples{
\dontrun{
my_triangle <- triangle(a = 3, b = 4, c = 5)
}

}
\seealso{
Other 2D Shape Classes: 
\code{\link{circle}()},
\code{\link{rectangle}()},
\code{\link{square}()}
}
\concept{2D Shape Classes}
