% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume.cube}
\alias{volume.cube}
\title{Calculate the volume of a cube.}
\usage{
\method{volume}{cube}(shp)
}
\arguments{
\item{shp}{An object of class '3dShape' representing a cube with side length.}
}
\value{
The volume of the cube.
}
\description{
Calculate the volume of a cube.
}
\examples{
\dontrun{
# Create a cube object
my_cube1 <- cube(sideLength = 5)

# Calculate the volume of the cube
volume.cube(my_cube1)
}
}
\seealso{
Other volume functions: 
\code{\link{volume.cuboid}()},
\code{\link{volume.cylinder}()},
\code{\link{volume.sphere}()},
\code{\link{volume}()}
}
\concept{volume functions}
