% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume.cuboid}
\alias{volume.cuboid}
\title{Calculate the volume of a cuboid.}
\usage{
\method{volume}{cuboid}(cu)
}
\arguments{
\item{cu}{An object of class '3dShape' representing a cuboid with dimensions
height, width, and depth.}
}
\value{
The volume of the cuboid.
}
\description{
Calculate the volume of a cuboid.
}
\examples{
# Create a cuboid object
my_cuboid <- cuboid(height = 3, width = 4, depth = 5)

# Calculate the volume of the cuboid
volume.cuboid(my_cuboid)

}
\seealso{
Other volume functions: 
\code{\link{volume.cube}()},
\code{\link{volume.cylinder}()},
\code{\link{volume.sphere}()},
\code{\link{volume}()}
}
\concept{volume functions}
