% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume.cylinder}
\alias{volume.cylinder}
\title{Calculate the volume of a cylinder.}
\usage{
\method{volume}{cylinder}(shp)
}
\arguments{
\item{shp}{An object of class '3dShape' representing a cylinder with radius and height.}
}
\value{
The volume of the cylinder.
}
\description{
Calculate the volume of a cylinder.
}
\examples{
# Create a cylinder object
my_cylinder <- cylinder(radius = 2, height = 5)

# Calculate the volume of the cylinder
volume.cylinder(my_cylinder)

}
\seealso{
Other volume functions: 
\code{\link{volume.cube}()},
\code{\link{volume.cuboid}()},
\code{\link{volume.sphere}()},
\code{\link{volume}()}
}
\concept{volume functions}
