% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume}
\alias{volume}
\title{Calculate the volume of 3D shapes.}
\usage{
volume(shp)
}
\arguments{
\item{shp}{An object of class '3dShape'.}
}
\description{
This provides functions to calculate the volume of various 3D shapes,
specifically cuboids, cubes, spheres, and cylinders. The volume can be calculated
based on the dimensions of each shape.
}
\details{
The function checks if the input object is of class '3dShape'. If not,
it raises an error indicating that the input must be an object of class '3dShape'.
If the input is of class '3dShape', it dispatches the calculation to the appropriate
method based on the type of shape.
}
\seealso{
Other volume functions: 
\code{\link{volume.cube}()},
\code{\link{volume.cuboid}()},
\code{\link{volume.cylinder}()},
\code{\link{volume.sphere}()}
}
\concept{volume functions}
