% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume.sphere}
\alias{volume.sphere}
\title{Calculate the volume of a sphere.}
\usage{
\method{volume}{sphere}(sph)
}
\arguments{
\item{sph}{An object of class '3dShape' representing a sphere with radius.}
}
\value{
The volume of the sphere.
}
\description{
Calculate the volume of a sphere.
}
\examples{
# Create a sphere object
my_sphere <- sphere(radius = 2)

# Calculate the volume of the sphere
volume.sphere(my_sphere)

}
\seealso{
Other volume functions: 
\code{\link{volume.cube}()},
\code{\link{volume.cuboid}()},
\code{\link{volume.cylinder}()},
\code{\link{volume}()}
}
\concept{volume functions}
