% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area.cuboid}
\alias{area.cuboid}
\title{Calculate the surface area of a cuboid.}
\usage{
\method{area}{cuboid}(cu)
}
\arguments{
\item{cu}{An object of class 'shape' representing a cuboid with dimensions depth, width, and height.}
}
\value{
The surface area of the cuboid.
}
\description{
Calculate the surface area of a cuboid.
}
\examples{
# Create a cuboid object
cuboid <- list(depth = 3, width = 4, height = 5)
class(cuboid) <- "shape"

# Calculate the surface area of the cuboid
area.cuboid(cuboid)

}
\seealso{
\code{\link{area}}

Other area functions: 
\code{\link{area.circle}()},
\code{\link{area.cube}()},
\code{\link{area.cylinder}()},
\code{\link{area.rectangle}()},
\code{\link{area.sphere}()},
\code{\link{area.square}()},
\code{\link{area.triangle}()},
\code{\link{area}()}
}
\concept{area functions}
