% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area.sphere}
\alias{area.sphere}
\title{Calculate the surface area of a sphere.}
\usage{
\method{area}{sphere}(sph)
}
\arguments{
\item{sph}{An object of class 'shape' representing a sphere with radius.}
}
\value{
The surface area of the sphere.
}
\description{
Calculate the surface area of a sphere.
}
\examples{
# Create a sphere object
sphere <- list(radius = 2)
class(sphere) <- "shape"

# Calculate the surface area of the sphere
area.sphere(sphere)
}
\seealso{
\code{\link{area}}

Other area functions: 
\code{\link{area.circle}()},
\code{\link{area.cube}()},
\code{\link{area.cuboid}()},
\code{\link{area.cylinder}()},
\code{\link{area.rectangle}()},
\code{\link{area.square}()},
\code{\link{area.triangle}()},
\code{\link{area}()}
}
\concept{area functions}
