% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area.square}
\alias{area.square}
\title{Calculate the area of a square.}
\usage{
\method{area}{square}(squ)
}
\arguments{
\item{squ}{An object of class 'shape' representing a square.}
}
\value{
The area of the square.
}
\description{
Calculate the area of a square.
}
\examples{
# Create a square object
square <- list(sideLength = 5)
class(square) <- "shape"

# Calculate the area of the square
area.square(square)


}
\seealso{
\code{\link{area}}

Other area functions: 
\code{\link{area.circle}()},
\code{\link{area.cube}()},
\code{\link{area.cuboid}()},
\code{\link{area.cylinder}()},
\code{\link{area.rectangle}()},
\code{\link{area.sphere}()},
\code{\link{area.triangle}()},
\code{\link{area}()}
}
\concept{area functions}
